backend/requirements.txt
fastapi[standard]
uvicorn 
langchain 
pytest 
python-dotenv 
sqlalchemy
langchain-community
openai
chromadb
langchain-openai
langchain-chroma
